requirements.txt
# ============================================================================
# ARIES PLATFORM - PYTHON DEPENDENCIES
# ============================================================================

# Core Data Processing
pyspark==3.5.0
pandas==2.0.3
numpy==1.24.3
polars==0.19.11

# Kafka & Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0

# Database
psycopg2-binary==2.9.9
sqlalchemy>=1.4.24,<2.0
pymongo==4.5.0

# NLP & Sentiment Analysis
vaderSentiment==3.3.2
textblob==0.17.1
scikit-learn==1.3.2
transformers==4.33.3  # For advanced models (BERT)
torch==2.0.1  # Required for transformers
nltk==3.8.1

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.14.0
requests==2.31.0

# Reddit API
praw==7.7.0
prawcore==2.4.0

# Cloud Storage
minio==7.2.0
boto3==1.29.7

# Web Framework
streamlit==1.31.1
plotly==5.17.0
altair==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
attrs==23.1.0
python-dateutil==2.8.2

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==1.37.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
mock==5.1.0

# Visualization & Reporting
matplotlib==3.8.0
seaborn==0.13.0
dash==2.14.1

# Development
black==23.11.0
flake8==6.1.0
pylint==3.0.2
isort==5.12.0
mypy==1.7.0

# Scheduling (Optional - for production)
apache-airflow==2.7.1
